+++ stranger/resources/__init__.py
@@ -13,7 +13,7 @@ repeats_json = "resources/variant_catalog_grch37.json"
 
 ###### Paths ######
 
-# Backround data path
+# Background data path
 
 repeats_path = files("stranger").joinpath(repeats_file)
 repeats_json_path = files("stranger").joinpath(repeats_json)